# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/produtos": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_produtos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProdutoLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Produtos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_produtos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/produtos/catalogo": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Criar Catalogo",
        "description": "Cria os produtos do catálogo do site (preços em branco, para a equipe definir).",
        "operationId": "criar_catalogo_api_v1_produtos_catalogo_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProdutoLeitura"
                  },
                  "type": "array",
                  "title": "Response Criar Catalogo Api V1 Produtos Catalogo Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/produtos/{id_}": {
      "put": {
        "tags": [
          "Produtos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_produtos__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_produtos__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/negocios": {
      "get": {
        "tags": [
          "Negócios"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_negocios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NegocioLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Negocios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Negócios"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_negocios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegocioEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegocioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/negocios/{id_}": {
      "put": {
        "tags": [
          "Negócios"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_negocios__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegocioAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegocioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Negócios"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_negocios__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/negocios/{id_}/etapa": {
      "patch": {
        "tags": [
          "Negócios"
        ],
        "summary": "Mover",
        "operationId": "mover_api_v1_negocios__id___etapa_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoverEtapa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegocioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/parceiros": {
      "get": {
        "tags": [
          "Parceiros de negócio"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_parceiros_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "papel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 40
                },
                {
                  "type": "null"
                }
              ],
              "title": "Papel"
            }
          },
          {
            "name": "busca",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParceiroLeitura"
                  },
                  "title": "Response Listar Api V1 Parceiros Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Parceiros de negócio"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_parceiros_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParceiroEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParceiroLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parceiros/{id_}": {
      "put": {
        "tags": [
          "Parceiros de negócio"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_parceiros__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParceiroAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParceiroLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Parceiros de negócio"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_parceiros__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parceiros/papeis": {
      "get": {
        "tags": [
          "Parceiros de negócio"
        ],
        "summary": "Papeis",
        "description": "Papéis que um parceiro pode assumir (qualquer pessoa logada pode consultar).",
        "operationId": "papeis_api_v1_parceiros_papeis_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PapelLeitura"
                  },
                  "type": "array",
                  "title": "Response Papeis Api V1 Parceiros Papeis Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoverEtapa": {
        "properties": {
          "etapa": {
            "type": "string",
            "enum": [
              "lead",
              "diag_agendado",
              "diag_feito",
              "proposta",
              "negociacao",
              "ganho",
              "perdido"
            ],
            "title": "Etapa"
          },
          "motivo_perda": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Preço",
                  "Prazo",
                  "Escolheu concorrente",
                  "Adiou o projeto",
                  "Sem resposta",
                  "Fora do perfil"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Perda"
          },
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "etapa",
          "versao"
        ],
        "title": "MoverEtapa"
      },
...
      "NegocioLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "etapa": {
            "type": "string",
            "title": "Etapa"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          },
          "mensal": {
            "type": "number",
            "title": "Mensal"
          },
          "previsao": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previsao"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "motivo_perda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Perda"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "fechado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fechado Em"
          },
          "faturado": {
            "type": "boolean",
            "title": "Faturado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "titulo",
          "cliente_id",
          "cliente_nome",
          "etapa",
          "valor",
          "mensal",
          "previsao",
          "responsavel_id",
          "origem",
          "motivo_perda",
          "obs",
          "fechado_em",
          "faturado"
        ],
        "title": "NegocioLeitura"
      },
...
      "PapelLeitura": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "title": "PapelLeitura"
      },
...
      "ProdutoAtualizar": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "projeto",
              "mensal",
              "consultoria",
              "outro"
            ],
            "title": "Tipo"
          },
          "unidade": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Unidade",
            "default": "projeto"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Preco",
            "default": "0"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "tipo",
          "versao"
        ],
        "title": "ProdutoAtualizar"
      },
      "ProdutoEntrada": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "projeto",
              "mensal",
              "consultoria",
              "outro"
            ],
            "title": "Tipo"
          },
          "unidade": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Unidade",
            "default": "projeto"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Preco",
            "default": "0"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "tipo"
        ],
        "title": "ProdutoEntrada"
      },
      "ProdutoLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "unidade": {
            "type": "string",
            "title": "Unidade"
          },
          "preco": {
            "type": "number",
            "title": "Preco"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "nome",
          "tipo",
          "unidade",
          "preco",
          "ativo",
          "descricao"
        ],
        "title": "ProdutoLeitura"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}